+++ path/src/lib.rs
@@ -48,7 +48,6 @@ mod events;
 mod path_state;
 mod path;
 pub mod generic;
-pub mod generic_4;
 pub mod polygon;
 pub mod iterator;
 pub mod builder;
